==> src/10_create_geometry_file.ipynb <==
# URL to download the 2023 county shapefiles
import geopandas as gpd

tiger2023_url = 'https://www2.census.gov/geo/tiger/TIGER2023/COUNTY/tl_2023_us_county.zip'
us_gdf = gpd.read_file(tiger2023_url)
us_gdf['GEOID'] = us_gdf['GEOID'].astype(str)
us_gdf.head()

import os
from zipfile import ZipFile

# Download the zip file from the provided URL
# https://data.humdata.org/dataset/9721eaf0-5663-4137-b3a2-c21dc8fac15a/resource/f151b1c1-1353-4f57-bdb2-b1b1c18a1fd1

folder_path = "../supporting_files"

mex_shp_file_path = os.path.join(folder_path, 'mex_admbnda_govmex_20210618_shp.zip')
# Define the extraction path
extraction_path = os.path.join(folder_path, 'mex_admbnda_govmex_20210618_shp')

# Unzip the file
with ZipFile(mex_shp_file_path, 'r') as zip_ref:
    zip_ref.extractall(extraction_path)

# List the contents of the extracted directory
extracted_files = os.listdir(extraction_path)
print("Extracted files:", extracted_files)

# Define the path to the nested extracted directory
nested_extraction_path = os.path.join(extraction_path, 'mex_admbnda_govmex_20210618_SHP')

# List the contents of the nested directory
nested_files = os.listdir(nested_extraction_path)
print("Nested directory files:", nested_files)


import os

# Assuming the correct shapefile is 'mex_admbnda_adm2_govmex_20210618.shp'
shapefile_name = 'mex_admbnda_adm2_govmex_20210618.shp'  # Replace with actual file name from the listing
shapefile_path = os.path.join(nested_extraction_path, shapefile_name)

# Load the shapefile into a GeoDataFrame
mx_gdf = gpd.read_file(shapefile_path)
mx_gdf.head()

# from mx_gdf retain only ADM2_PCODE and geometry columns
mx_gdf = mx_gdf[['ADM2_PCODE', 'geometry']]
us_gdf = us_gdf[["GEOID", "geometry"]]

# rename ADM2_PCODE to COUNTY_MUNI_CODE in mx_gdf
mx_gdf = mx_gdf.rename(columns={"ADM2_PCODE": "COUNTY_MUNI_CODE"})
# zfill to ensure 5-digit codes and add prefix 'US
us_gdf["GEOID"] = us_gdf["GEOID"].str.zfill(5)
us_gdf["GEOID"] = "US" + us_gdf["GEOID"]
us_gdf = us_gdf.rename(columns={"GEOID": "COUNTY_MUNI_CODE"})
us_gdf.head()

mx_gdf.head()

us_gdf = us_gdf.to_crs(epsg=4326)
mx_gdf = mx_gdf.to_crs(epsg=4326)

import pandas as pd

combined_gdf = pd.concat([us_gdf, mx_gdf], ignore_index=True)
combined_gdf = gpd.GeoDataFrame(combined_gdf, geometry='geometry', crs='EPSG:4326')
combined_gdf.shape

import os

output_csv_path = '../generated_files/geometry_file.csv'
os.makedirs(os.path.dirname(output_csv_path), exist_ok=True)
combined_gdf.to_csv(output_csv_path, index=False)

output_shapefile_path = '../generated_files/us_mx_combined_shapefile.shp'
combined_gdf.to_file(output_shapefile_path)

# Summarize in a sentence what was done in this code.
# This code downloads and processes US and Mexico county shapefiles, combines them into a single GeoDataFrame, and saves the result as both a CSV and a shapefile.
