==> src/20_calculate_centroid.ipynb <==
import geopandas as gpd

output_shapefile_path = '../generated_files/us_mx_combined_shapefile.shp'
# load shape file
combined_gdf = gpd.read_file(output_shapefile_path)

combined_gdf.columns

combined_gdf = combined_gdf.set_crs(epsg=4326)
combined_gdf = combined_gdf.to_crs(epsg=3857)
combined_gdf['centroid'] = combined_gdf.geometry.centroid
centroids_gdf = gpd.GeoDataFrame(combined_gdf[['COUNTY_MUN']], geometry=combined_gdf['centroid'], crs='EPSG:3857')

centroids_gdf.head()

centroids_gdf.rename(columns={'COUNTY_MUN': 'COUNTY_MUNI_CODE'}, inplace=True)
centroids_gdf = centroids_gdf.to_crs(epsg=4326)
# Extract latitude and longitude
centroids_gdf['Latitude'] = centroids_gdf.geometry.y
centroids_gdf['Longitude'] = centroids_gdf.geometry.x
centroids_gdf.head()

cols_to_retain = ['COUNTY_MUNI_CODE', 'Latitude', 'Longitude']
centroids_final = centroids_gdf[cols_to_retain]
centroids_final.head()

centroids_final.to_csv('../generated_files/centroids_us_mx.csv', index=False)
