==> src/30_create_base_file.ipynb <==
import pandas as pd

county_names_file = "../original_data/all_counties_list.txt"

df_base = pd.read_csv(county_names_file, names=["COUNTY_MUNI_CODE"])
df_base.head()

all_centroid_file = '../generated_files/centroids_us_mx.csv'
df_centroids = pd.read_csv(all_centroid_file)
df_centroids.head()

row1 = ["MX04013", "20.3198", "-90.0568"]
row2 = ["MX07122", "16.3374", "-92.9744"]
comb_data = [row1, row2]
df_new = pd.DataFrame(comb_data, columns=["COUNTY_MUNI_CODE", "Latitude", "Longitude"])
df_centroids = pd.concat([df_centroids, df_new], ignore_index=True)
df_centroids.head()

# counties not in centroid file
counties_base = set(df_base['COUNTY_MUNI_CODE'])
counties_centroids = set(df_centroids['COUNTY_MUNI_CODE'])
missing_counties = counties_base - counties_centroids
missing_counties

df_base = df_centroids[df_centroids['COUNTY_MUNI_CODE'].isin(df_base['COUNTY_MUNI_CODE'])]

df_base

### Finding nearest latitude and longitude according to .nc file for speed

df_base["nearest_latitude"] = None
df_base["nearest_longitude"] = None

import xarray as xr

path = "../ncs/r_250_mon_ERA5_1979-2025.nc"  # adjust if your file is elsewhere
ds = xr.open_dataset(path, decode_times=True)
ds

all_lats = ds.latitude.values
all_longs = ds.longitude.values

import numpy as np
from scipy.spatial import cKDTree
from math import radians, cos

def build_kdtree(lat_grid, lon_grid):
    """
    Build a KDTree from 2D lat-lon grid (145x236).
    Returns:
        tree: cKDTree object
        points: array of shape (145*236, 2) containing all (lat, lon) pairs
    """
    # Create full coordinate mesh
    lat_mesh, lon_mesh = np.meshgrid(lat_grid, lon_grid, indexing='ij')
    points = np.column_stack((lat_mesh.ravel(), lon_mesh.ravel()))
    tree = cKDTree(points)
    return tree, points

tree, points = build_kdtree(all_lats, all_longs)

def find_nearest_point(lat_c, lon_c, tree, points):
    """
    Finds the nearest (lat, lon) from the KDTree to the county center.
    Returns: nearest_lat, nearest_lon, distance_deg
    """
    dist, idx = tree.query([lat_c, lon_c])
    nearest_lat, nearest_lon = points[idx]
    return nearest_lat, nearest_lon, dist

center_lat_sample = 22.9913638229323
center_lon_sample = -102.475691706378
nearest_lat, nearest_lon, dist = find_nearest_point(center_lat_sample, center_lon_sample, tree, points)
nearest_lat, nearest_lon, dist

for i in range(len(df_base)):
    center_lat_sample = df_base.iloc[i]["Latitude"]
    center_lon_sample = df_base.iloc[i]["Longitude"]
    nearest_lat, nearest_lon, dist = find_nearest_point(center_lat_sample, center_lon_sample, tree, points)
    df_base.loc[df_base.index[i], "nearest_latitude"] = nearest_lat
    df_base.loc[df_base.index[i], "nearest_longitude"] = nearest_lon
df_base.head()

import numpy as np

num_unique_counties = len(df_base)
all_dates = pd.date_range(start="1979-01-01", end="2025-08-01", freq="MS")
print(len(all_dates))
df_monthly = pd.concat([df_base]*len(all_dates))
df_monthly = df_monthly.sort_index()
df_monthly.head()


df_monthly['date'] = np.tile(all_dates, num_unique_counties)
df_monthly

col = 'date'
df_monthly = df_monthly[[col]+[c for c in df_monthly.columns if c!=col]]
df_monthly

df_monthly.reset_index(drop=True, inplace=True)
df_monthly

import pickle
import os

os.makedirs('../main_data', exist_ok=True)

with open('../main_data/base_file.pkl', 'wb') as f:
    pickle.dump(df_monthly, f)
