==> src/40_combine_all.ipynb <==
import os
import pickle

base_folder = "../aux_data"

files = os.listdir(base_folder)
files = [f for f in files if f.endswith(".pkl")]
files = [f for f in files if f.startswith("df_monthly_")]
# files = [f for f in files if "new_counties" not in f]
files.sort()

# Load the first file as the base dataframe
with open(os.path.join(base_folder, files[0]), 'rb') as f:
    df_final = pickle.load(f)

df_final.head()

import pandas as pd

# Merge progressively to save memory
for fname in files[1:]:
    file_path = os.path.join(base_folder, fname)
    with open(file_path, 'rb') as f:
        df_next = pickle.load(f)

    # Find common columns to merge on
    common_cols = list(set(df_final.columns) & set(df_next.columns))
    if not common_cols:
        raise ValueError(f"No common columns between {fname} and current df.")

    # Merge on the common columns
    df_final = pd.merge(df_final, df_next, on=common_cols, how='outer')

    # Free memory
    del df_next

out_folder ="../main_data/"
with open(os.path.join(out_folder, "combined_v2.pkl"), "wb") as f:
    pickle.dump(df_final, f)

df_final
