==> src/60_datetime_to_PII.ipynb <==
import pandas as pd

pii_file = "../generated_files/vsv_county_noPII_coded.csv"
df_pii = pd.read_csv(pii_file)
df_pii.head()

from datetime import datetime

years = df_pii.ONSET_YEAR.tolist()
months = df_pii.ONSET_MONTH.tolist()

datetimes = []
for year, month in zip(years, months):
    datetimes.append(datetime(year, int(month), 1))

df_pii["TIME"] = datetimes
df_pii.head()

df_pii.drop(columns=["ONSET_MONTH", "ONSET_DAY"], inplace=True)
# Move TIME to the first column
df_pii = df_pii[["TIME"] + [col for col in df_pii.columns if col != "TIME"]]
df_pii.head()

df_pii.to_csv("../generated_files/vsv_county_noPII_coded_v2.csv", index=False)
