==> src/70_integrate_cases_w_feats.ipynb <==
import pickle

monthly_file = "../main_data/combined_v2.pkl"
with open(monthly_file, "rb") as f:
    df_feats = pickle.load(f)

df_feats.head()

df_feats["cases"] = None
df_feats

import pandas as pd

pii_file = "../generated_files/vsv_county_noPII_coded_v2.csv"
pii_df = pd.read_csv(pii_file)
pii_df.head()

df_pii_cases = (
    pii_df.groupby(["COUNTY_MUNI_CODE", "TIME"])
          .size()
          .reset_index(name='NUM_CASES')
)
df_pii_cases

import warnings
warnings.filterwarnings("ignore")

new_county_case = []
for i in range(len(df_pii_cases)):
    if i % 100 == 0:
        print(f"Processing {i} out of {len(df_pii_cases)}")
    cmc = df_pii_cases.iloc[i]["COUNTY_MUNI_CODE"]
    time = df_pii_cases.iloc[i]["TIME"]
    num_cases = df_pii_cases.iloc[i]["NUM_CASES"]
    
    try:
        index_in_monthly = df_feats[df_feats["COUNTY_MUNI_CODE"] == cmc][df_feats["date"] == time].index[0]
        df_feats.at[index_in_monthly, "cases"] = num_cases
    except IndexError:
        print(f"IndexError for CMC: {cmc}, TIME: {time}")
        pass

df_feats.isnull().sum()

df_feats = df_feats.fillna(0)
df_feats.isnull().sum()

with open("../main_data/combined_v3.pkl", "wb") as f:
    pickle.dump(df_feats, f)
