==> src/80_integrate_sp_ser.ipynb <==
import pickle

with open("../main_data/combined_v3.pkl", "rb") as f:
    df = pickle.load(f)

df.head()

import pandas as pd

pii_file = "../generated_files/vsv_county_noPII_coded_v2.csv"
pii_df = pd.read_csv(pii_file)
pii_df.head()

new_sp_columns = ["ser_IN", "ser_NJ", "ser_unknown", "sp_bovine", "sp_camelid", "sp_caprine", "sp_equine", "sp_ovine", "sp_porcine", "sp_suilline", "sp_rhinocerus", "sp_unknown"]



for col in new_sp_columns:
    df[col] = 0

df.head()

sp_dict = {"bovine": "sp_bovine",
            "camelid-alpaca": "sp_camelid",
            "camelid-llama": "sp_camelid",
            "caprine": "sp_caprine",
            "equine": "sp_equine", 
            "ovine": "sp_ovine",
            "porcine": "sp_porcine",
            "suilline": "sp_suilline",
            "rhinoceros": "sp_rhinocerus",
            "nan": "sp_unknown"
            }
ser_dict = { "IN": "ser_IN",
             "NJ": "ser_NJ",
             "nan": "ser_unknown"
             }

for i in range(len(pii_df)):
    if i % 10 == 0:
        print(f"Processing row {i} of {len(pii_df)}")
    time = pii_df.loc[i, "TIME"]
    county_muni_code = pii_df.loc[i, "COUNTY_MUNI_CODE"]
    ser = str(pii_df.loc[i, "SEROTYPE"])
    species = str(pii_df.loc[i, "SPECIES"])

    if species == "bovine and equine":
        index_num = df.loc[
        (df["date"] == time) &
        (df["COUNTY_MUNI_CODE"] == county_muni_code),
        ].index.values[0]
        df.at[index_num, "sp_bovine"] += 1
        df.at[index_num, "sp_equine"] += 1
        continue
    
    alloted_col_ser = ser_dict[ser]
    alloted_col_sp = sp_dict[species]
    try:
        index_num = df.loc[
        (df["date"] == time) &
        (df["COUNTY_MUNI_CODE"] == county_muni_code)
        ].index.values[0]
        df.at[index_num, alloted_col_ser] += 1
        df.at[index_num, alloted_col_sp] += 1
    except:
        print(f"Error at time: {time} and county_muni_code: {county_muni_code}, ser: {ser}, species: {species}")

out_file = "../main_data/VSV_dynamic_data.pkl"
with open(out_file, "wb") as f:
    pickle.dump(df, f)
